[chore] RemainingCountBadgeView 파일 분리
